github-actions: attempt using prebuilt Docker image
*This is a test!*

We currently build the Docker image once per training, for all commits
or PRs on the repository. This is slow, at least 6m of image build time
per training (there are 11). This amounts to 1h19m for a random commit.

Let's switch to a prebuilt image on Docker Hub. We cannot use Github
Registry has storage limitations (even on enterprise accounts).

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>